refactor: Enable ruff and ament_mypy checks and fix lint errors (#1063)
* Enable flake8-bugbear and fix lint errors

* Enable flake8-builtins and fix lint errors

* Enable flake8-datetimez and fix lint errors

* Enable flake8-executable and fix lint errors

* Fix lint error in rosapi_node

* Enable flake8-future-annotations and fix lint errors

* Enable flake8-implicit-str-concat and fix lint errors

* Enable flake8-pie and fix lint errors

* Enable flake8-raise and fix lint errors

* Enable flake8-return and fix lint errors

* Enable flake8-simplify and fix lint errors

* Enable flake8-type-checking and fix lint errors

* Enable flake8-unused-arguments and fix lint errors

* Enable flake8-use-pathlib and fix lint errors

* Enable Perflint and fix lint errors

* Enable pyupgrade and fix lint errors

* Enable refurb and fix lint errors

* Enable ruff-specific checks and fix errors

* Fix subscribe test

* Fix mypy errors

* rosapi: type check test methods

* Fix rosapi_node lint errors

* Set max line length to 120 and fix some of the errors

* Fix the rest of 'line too long' errors

* Don't change argument names in public API methods

* Enable flake8-errmsg and fix lint errors

* Enable flake8-logging and fix lint error

* Enable flake8-pyi and fix lint errors

* Fix pylint errors

* Add comments to ruff config

* Fix trailing whitespace

* Enable flake8-logging-format

* Simplify message handler transitions

* Fix other ruff lint errors

* Add ament_mypy test and fix mypy errors

* Fix ruff check in rosapi_node

+++ rosbridge_library/src/rosbridge_library/capabilities/fragmentation.py
@@ -90,7 +90,7 @@ class Fragmentation(Capability):
 
         log_msg = (
             "sending "
-            + str(int(math.ceil(message_length / float(fragment_size))))
+            + str(math.ceil(message_length / float(fragment_size)))
             + " parts [fragment size: "
             + str(fragment_size)
             + "; expected duration: ~"
